+++ zed_components/src/zed_camera/include/sl_types.hpp
@@ -44,6 +44,7 @@
 #include <sensor_msgs/msg/nav_sat_status.hpp>
 #include <sensor_msgs/msg/point_cloud2.hpp>
 #include <sensor_msgs/msg/temperature.hpp>
+#include <std_msgs/msg/bool.hpp>
 #include <std_srvs/srv/set_bool.hpp>
 #include <std_srvs/srv/trigger.hpp>
 #include <stereo_msgs/msg/disparity_image.hpp>
@@ -134,6 +135,8 @@ typedef std::shared_ptr<rclcpp::Publisher<geographic_msgs::msg::GeoPoseStamped>>
   geoPosePub;
 typedef std::shared_ptr<rclcpp::Publisher<sensor_msgs::msg::NavSatFix>>
   gnssFixPub;
+
+typedef std::shared_ptr<rclcpp::Publisher<std_msgs::msg::Bool>> healthPub;
 
 typedef std::shared_ptr<rclcpp::Subscription<geometry_msgs::msg::PointStamped>> clickedPtSub;
 typedef std::shared_ptr<rclcpp::Subscription<sensor_msgs::msg::NavSatFix>> gnssFixSub;